requirements.txt
google-cloud-aiplatform[adk,agent-engines]>=1.93.0,<2.0.0
google-genai>=1.9.0,<2.0.0
pydantic>=2.10.6,<3.0.0
python-dotenv>=1.0.1,<2.0.0
google-adk>=1.0.0,<2.0.0
requests>=2.31.0,<3.0.0
google-adk[eval]
pytest
pytest_asyncio
